Add DeleteButton
