+++ src/usapyon_ui.js
@@ -21,19 +21,6 @@ function stopAll() {
   document.getElementById("sound8").pause();
 }
 
-$("#del").click(function() {
-  if (localStorage.usapri == 1) {
-    alert("消せません");
-  } else {
-    var del = confirm("全てのデータを初期化します。よろしいですか？");
-    if (del) {
-      localStorage.clear();
-      clearInterval(setInterval(achievement, 10));
-      location.reload();
-    }
-  }
-});
-
 $("#version").html("ver.1.1.1β");
 
 $("#version").click(function() {
